refactor: account service to logic directory
Took 5 minutes
