+++ src/main/java/dev/realtards/kuenyawz/services/AuthServiceImpl.java
@@ -9,6 +9,7 @@ import dev.realtards.kuenyawz.entities.RefreshToken;
 import dev.realtards.kuenyawz.exceptions.InvalidPasswordException;
 import dev.realtards.kuenyawz.exceptions.UnauthorizedException;
 import dev.realtards.kuenyawz.mapper.AccountMapper;
+import dev.realtards.kuenyawz.services.entity.AccountService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
