+++ src/test/java/dev/realtards/kuenyawz/controllers/AccountControllerIT.java
@@ -3,9 +3,8 @@ package dev.realtards.kuenyawz.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dev.realtards.kuenyawz.boostrappers.DatabaseBootstrapper;
 import dev.realtards.kuenyawz.dtos.account.AccountSecureDto;
-import dev.realtards.kuenyawz.entities.Account;
 import dev.realtards.kuenyawz.repositories.AccountRepository;
-import dev.realtards.kuenyawz.services.AccountService;
+import dev.realtards.kuenyawz.services.entity.AccountService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
